Switch time_t, ino_t, clock_t, and struct kevent's ident and data members to 64bit types.  Assign new syscall numbers for (almost all) the syscalls that involve the affected types, including anything with time_t, timeval, itimerval, timespec, rusage, dirent, stat, or kevent arguments.  Add a d_off member to struct dirent and replace getdirentries() with getdents(), thus immensely simplifying and accelerating telldir/seekdir.  Build perl with -DBIG_TIME.
Bump the major on every single base library: the compat bits included
here are only good enough to make the transition; the T32 compat
option will be burned as soon as we've reached the new world are
are happy with the snapshots for all architectures.

DANGER: ABI incompatibility.  Updating to this kernel requires extra
work or you won't be able to login: install a snapshot instead.

Much assistance in fixing userland issues from deraadt@ and tedu@
and build assistance from todd@ and otto@

+++ sys/sys/sysctl.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: sysctl.h,v 1.136 2013/06/09 13:10:19 miod Exp $	*/
+/*	$OpenBSD: sysctl.h,v 1.137 2013/08/13 05:52:27 guenther Exp $	*/
 /*	$NetBSD: sysctl.h,v 1.16 1996/04/09 20:55:36 cgd Exp $	*/
 
 /*
@@ -398,7 +398,7 @@ struct kinfo_proc {
 
 	int64_t	p_uvalid;		/* CHAR: following p_u* members from struct user are valid */
 					/* XXX 64 bits for alignment */
-	u_int32_t p_ustart_sec;		/* STRUCT TIMEVAL: starting time. */
+	u_int64_t p_ustart_sec;		/* STRUCT TIMEVAL: starting time. */
 	u_int32_t p_ustart_usec;	/* STRUCT TIMEVAL: starting time. */
 
 	u_int32_t p_uutime_sec;		/* STRUCT TIMEVAL: user time. */
